requirements.txt
# BorrowHood v1.0 Dependencies
# Pin versions for reproducible builds

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Auth (JWT)
python-jose[cryptography]==3.3.0
httpx==0.28.1

# Validation & config
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Templates
jinja2==3.1.5

# Utilities
python-slugify==8.0.4
python-multipart==0.0.20

# Telegram notifications
# (uses httpx -- already listed above)
